Keep ledger summaries in scope
